requirements.txt
numpy
scipy
scikit-image
bd-metric

opencv-python
pandas
PyQt6
matplotlib
XlsxWriter
openpyxl
pyjanitor